requirements.txt
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.14.2
billiard==4.2.1
black==24.4.2
celery==5.3.6
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.5.1
cryptography==42.0.5
curl_cffi==0.13.0
distlib==0.4.0
Django==5.0.6
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
factory-boy==3.3.0
Faker==37.5.3
filelock==3.18.0
flake8==7.0.0
frozendict==2.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
identify==2.6.13
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==5.13.2
joblib==1.5.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
mccabe==0.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.5
nvidia-nccl-cu12==2.28.9
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
peewee==3.18.3
pillow==10.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre-commit==3.7.1
prompt_toolkit==3.0.51
protobuf==6.33.1
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.4
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
scikit-learn==1.7.2
scipy==1.16.3
sentry-sdk==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.33.1
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.6.0
xgboost==3.1.2
yfinance==0.2.66
